Added functions for instant transfer, ratio slider and fixed bugs
Two new functions were added to handle the ratio slider which was set up as a listener in order to make them as cross browser and mobile friendly as possible. Another funtion was added to redirect to steem connect for instant transfer. Also fixed a number of bugs referenced by issues 6,7,8,9.