Added logic for invite count and ajax
The login method was updated to include the pending invite count in the page template and the invite method was updated to include logic for dealing with an ajax request. A flag was used to indicate whether the equest was ajax or not, allowing non javascript browsers to fall back to a regular html request.