Added better error handling for invalid ratios
Also adjusted ratio slider so that it has three layers of granularity. AT the furthest left it's range is 0.01 to 1, in the middle the range is from 1.0 to 9.9 and finally the far right's range is from 10 to 100 (whole integers). This was to help find optimal ratios with users who have small upvote values.