Created CSS for improved table displayed on info and archive page.
As per the mockup given by @camiloferrua (see documentation folder) changed the GUI for the info and archive tables, which now include the clock icon and the blue exchange icon. Buttons for accepting, canceling, bartering and sending invites are now all color coded so that when the reveal box is opened the Steem Connect button makes more sense.